Create Is Subsequence.py
New concepts learnt:

--> Split method (even though it was not used in the code, I tried coding with it)
Split method takes in two arguments, the regular expression and the string on which the expression applies. Using that regex, a list is generated. 

I came across an error while using spilt: TypeError: split() missing 1 required positional argument: 'string'
This is how I learnt about the arguments. 

-->Use of pointer
A pointer variable was used, it incremented every time a letter from S string matched with the T string letter. The pointer acted as a record book that counted the occurrence of the letter in string T. This way, without even creating another list, we were able to decide whether a subsequence existed or not.